Ajustes.

+++ include/continental/hydrotools/HeuristicSinkRemovalUtil.h
@@ -9,7 +9,7 @@ class HeuristicSinkRemovalUtil
 {
     public:
         //Retorna o indicador da direção de fluxo comparando duas células adjacentes
-        static short relativeIncipientFlowDirection(size_t x1, size_t x2, size_t y1, size_t y2);
+        static short relativeIncipientFlowDirection(int x1, int x2, int y1, int y2);
 
         //Move a linha e a coluna de acordo com o número armazenado
         static void moveToFlowDirection(short value, int &row, int &column);
